refactor: 카드 등록 화면 네이밍 변경
