+++ app/src/main/java/woowacourse/payments/ui/cardform/component/CardExpirationDateTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.registration.component
+package woowacourse.payments.ui.cardform.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
